+++ ydb/core/tx/datashard/change_record_cdc_serializer.cpp
@@ -1,6 +1,6 @@
 #include "change_record_cdc_serializer.h"
 #include "change_record.h"
-#include "export_common.h"
+#include "type_serialization.h"
 
 #include <ydb/core/protos/change_exchange.pb.h>
 #include <ydb/core/protos/grpc_pq_old.pb.h>
@@ -9,6 +9,7 @@
 #include <ydb/library/binary_json/read.h>
 #include <ydb/library/uuid/uuid.h>
 #include <ydb/library/yverify_stream/yverify_stream.h>
+#include <ydb/public/api/protos/ydb_topic.pb.h>
 
 #include <library/cpp/digest/md5/md5.h>
 #include <library/cpp/json/json_reader.h>
