+++ ydb/core/tx/datashard/export_s3_buffer_raw.cpp
@@ -1,6 +1,6 @@
 #ifndef KIKIMR_DISABLE_S3_OPS
 
-#include "export_common.h"
+#include "type_serialization.h"
 #include "export_s3_buffer_raw.h"
 
 #include <ydb/core/tablet_flat/flat_row_state.h>
